Amend instructions for connecting to host services
The IP of `docker0` needs to be provided to the container for it to be able to connect to e.g. Postgres. One way is the `--add-host` flag for `docker run`. This change adds instructions for doing so, so that people don't need to go looking for this info themselves.