update executor types with link to ios-tutorial
Maybe we need a complete section here about the macos executor type with information about xcode versions, etc? I just realized `macos` is really a third executor type in 2.0 and want to be sure customers understand that difference and the benefits of it being parallel to `machine` and `docker` in the config.