Warn $CIRCLE_WORKING_DIRECTORY not subbed in YAML
The prior language suggested "you" could use this. And since this is part of the configuration file docs, I used it in my config file. And it of course failed to find anything in the precise path including the dollar sign and all.

This change gives the specific context in which that environment variable will be available, and calls out explicitly that that context is not paths in the YAML config file.

This misunderstanding has tripped up others: https://discuss.circleci.com/t/cant-access-env-variables-from-my-yml-file/5463