move local dev instructions to local dev doc
Clean up CONTRIBUTING.md to focus on issues & pull requests. The very
Jekyll-specific stuff could arguably be removed from the local dev doc,
but we'll save that for a future PR.

This change should encourage more issues and fewer PRs -- the preferred
method of receiving and incorporating feedback into the docs.