Document CIRCLE_WORKFLOW_ID in env-vars.md
I wanted a way to consistently name artifacts created by one job that could be referenced by another job (i.e. for tagging Docker images). It took me a long time to realize that the undocumented `CIRCLE_WORKFLOW_ID` env var was available to me. 

I don't know if `CIRCLE_WORKFLOW_ID` is an "official" env var, and whether the description is accurate or not.