Document disabling automatic garbage collection
We are disabling automatic garbage collection because it is unnecessary in most
cases and in some cases (when it runs in the background for performance reasons)
can cause a cached `.git` directory to be corrupted.

Documenting this in caching.md (under Source Caching) and in
configuration-reference.md (for the docs on the `checkout` step itself).