Update .ruby-version to reflect version in Gemfile (#3492)
Currently the project cannot be `bundle install`ed because the Gemfile
specifies a different version of ruby than a ruby version manager would
have selected based on the .ruby-version file. Trying to results in an
error about the version mismatch.

The ruby version specified in the gemfile was recently updated in
9faeda7, but the version specified in `.ruby-version` (added in
3ce0a83) was not updated to reflect that.

We should either update .ruby-version or remove that file entirely and
let tools read it out of the Gemfile instead.