Update macOS Docs
The shell is now always a login shell, so we remove the instructions to
add --login on the 'testing on ios' page.

We also need to make this clear on the config references page.

Add docs for Xcode 11.1 and update the 11.0 docs to remove the "GM"
designation, since GM 2 _was_ the final release.