Provide a way to pass config data from Jekyll to JS
There are cases where we need to expose data that we have in the
Ruby/Jekyll context to the JS context.

This enables that by building a `circleJsConfig` JS object in an
HTML/JS snippet that gets included on every page.

In this initial usage, we use it to pass Algolia config data stored in
`_config.yml` to instantsearch.js.