[CIRCLE-23312] Make global nav links language-aware
Since this project is not fully internationalized, we don't have an
easy way to generate language-relative URLs for linking within a
language site.

We can work around that though by using the existing `lang` param on
pages that are part of non-English page collections.

This adds a plugin and Liquid filter to generate a URL to "outer" or
the circleci.com site, and employs it on links to outer in the global
nav include.

https://circleci.atlassian.net/browse/CIRCLE-23312