Handle language-relative URLs within docs project
This is a followup to the previous commit that produces
language-relative URLs to the outer (circleci.com) project.

This adds a second Liquid filter `language_relative_url` for creating
those based on the language of the current page, and also extracts
common page language prefix logic to a separate module called by both
this and the first `outer_url` filter.