Make nav bar logo tab-focusable
We want the full nav menu to be keyboard-navigable, but the logo did
not appear to be because the logo image element was being absolutely
positioned, causing the enclosing anchor tag to be sized 0x0.

Arguably this should not be absolutely positioned at all, however
fixing that is outside the purview of this ticket.

To work around that and get the anchor to be correctly sized, we shift
the positioning styles to the anchor rather than the image, leaving
only the height-defining rule on the image itself.

This ports over similar changes from circleci.com/outer:
https://github.com/circleci/circleci.com/pull/3030/