[CIRCLE-24660] Fix search box close button
This appears to be broken on Algolia's side in the widget, as the close
button is their native element that they create when the widget is
created. I assume this originally had the timeout to allow Algolia to
do its reset, which hopefully is completed by 100ms later when we
trigger a re-render with the now-empty search field.

However this isn't working for whatever reason. Instead we can do what
blog search does and clear the query string ourselves and immediately
re-render. It's nice to be able to remove that timeout/asynchronous
behavior too.