security-server: Add support bundle processing
Make customers aware that we upload traces from support bundles to a
third-party vendor, what kind of data they contain, and how long they
are retained for. This is based on functionality that was shipped in
2.19.3 and the process in our internal documentation.

It's not strictly an integration because the data is "manually" uploaded
after we receive a support bundle from a customer, but this page and
heading seems like the most appropriate place to put it.