Use snippet as workaround for broken internal link
It seems we translated against old version of the page that doesn't
use snippet and the links are broken. While we can fix this by
syncing Crowding to use the latest source file, I'm afraid that
enabling sync will mess up existing trnaslation, so we will be using
this workaround for now.