Update lang from "run" to assume Workflow usage
Workflows are now active on all projects on CircleCI. There is no longer a "Run", which may now be considered a pipeline. One piece of information omitted in this commit that might be worth adding back, if the user does not specify a workflow, and has a job named "build" we will automatically generate a workflow for the build job.