Finish InstantSearch.js upgrade
This incorporates API and config changes for upgrading from v2 to v4 of
InstantSearch and the Algolia client.

The only significant departure from the way it works before is removing
the lodash-based debounce function, which wasn't _not_ working, but it
wasn't working quite right. Something about it was causing the browser
to drop keystrokes from the query, very frustrating.

I tried just removing it and that verified it was the cause of that
issue, and since I think it feels/works better not-debounced, I decided
to just leave it out. Algolia discusses debouncing in their docs, and
they only recommend it for slow connections, so it seems like it isn't
intended to be debounced under normal circumstances.