Remove meaningless and unused environment variable from build job
This variable was originally introduced in 0fd59b4a in 2016-03:
https://github.com/circleci/circleci-docs/commit/0fd59b4a0a298c9d95e9ecd

On migration from v1 to v2, this project did not remove the variable:
https://github.com/circleci/circleci-docs/pull/1048

Ref. Jekyll is not using the variable any more with CircleCI 2.0
https://github.com/jekyll/jekyll/pull/8410

Also this variable has been set only for `build` job, but `ruby-deps`
step is used in other jobs, which is inconsistent with the job.

Signed-off-by: Takuya Noguchi <takninnovationresearch@gmail.com>