Update subnav shared asset, fire event on it when we show it
There was a bug in subnav.js due to it not being visible upon page
load, which is when it calculates the active anchor and subnav position
(yields incorrect results because the positions of all hidden elements
are 0).

This updates subnav.js to listen for a `shown.subnav` event and refresh
itself.

Then we fire that event upon showing search results including subnav.