[CIRCLE-31528] Align search results with search bar and nav
This applies a design change to align the left side of search results
with the left side of the search bar spyglass and header/nav.
Previously it was indented.

I didn't un-indent the contents of the div that was removed in order to
keep the diff clean and not muddy the line history, but I'm happy to
change that if desired.

Note that using a `row` element without a child `col` element is a
misuse of the Bootstrap grid system, but the header and search bar are
already doing this so we have to make the same mistake here to get them
to align with each other.