Correct path to packages folder in Yarn 1 example
The current suggested path to cache in the Yarn 1.x example actually points to the Yarn 2 install cache folder, making this a rather misleading example. Yarn 1.x creates a project level `node_modules` folder when installing, in the same way as NPM does. The path `./node_modules` is much more likely to point to the correct folder.