[CIRCLE-32372] Implement tabs for search result categories
Previously this was a scrollable list with the subnav tabs acting to
scroll the list to the given section of results, similar to how subnavs
work on circleci.com.

However this had some usability problems in this context and behaved
unexpectedly given the tab-looking nature of the subnav component.

This replaces that with Bootstrap's tab.js component.