Add custom Liquid tag for including a localized template
Laying the groundwork for internationalizing more of the site, this
adds an `include_localized` Liquid tag paralleling the normal
`include`, but will attempt to include a template from a language-code
subdirectory if the page is in a non-default language.

Example usage, replace this:

```
{% include global-nav.html %}
```

with:

```
{% include_localized global-nav.html %}
```

and Jekyll will render `_includes/ja/global-nav.html` on a
Japanese-language page and the default `_includes/global-nav.html` on
an English-language page.