Clarify that api v1 /build endpoint is not available on Server
Description

Clearly state that `https://circleci.com/api/v1.1/project/:vcs-type/:username/:project/build` endpoint is not available for CircleCI Server.

Reasons

To use this endpoint, one must enable pipelines, but pipeline is a CircleCI 2.1 feature and it is not available for CircleCI Server. Since it is not stated clearly, it is causing some confusion.