Make search result category tabs sticky
We can use the `dynamic-fixed` mode of the subnav component to
accomplish this, with a custom top nav offset param to account for the
fixed-position search bar.

We also need to scroll the document back to the top upon tab changes,
which we can do by attaching a handler to the boostrap tab shown event.