Language selector now redirects to the same content (#5110)
Make sidebar language selector to redirect to the same content.

\# Background

It used to redirect to base url. So, I was not using this selector,
because directly writing `ja` to the URL were much easier.

\# Description

This implementation relies on regexp to parse the current full URL to
replace the `docs/ja` or `docs` part to the target slug.