Update server-3-install.adoc (#5356)
* Update server-3-install.adoc

Moved nomad mtls instructions down in the documentation as it was causing people to install nomad.

* Update server-3-operator-backup-and-restore.adoc

Add note that this isn't available yet.

* Update server-3-install.adoc

This section is really not ideal. I'd like to make the mTLS default to disabled so we can remove this entire section and tell the user to ignore till step 4. For now this is a slight improvement to provide more context.

* Adding links

Co-authored-by: annabaker <anna.baker.313@gmail.com>