Changes to Server 3 Ops/Metrics Guide (#5403)
* Update server-3-install.adoc

Moved nomad mtls instructions down in the documentation as it was causing people to install nomad.

* Update server-3-operator-backup-and-restore.adoc

Add note that this isn't available yet.

* Update server-3-install.adoc

This section is really not ideal. I'd like to make the mTLS default to disabled so we can remove this entire section and tell the user to ignore till step 4. For now this is a slight improvement to provide more context.

* Adding links

* Update server-3-operator-metrics-and-monitoring.adoc

We removed LOKI and Grafana from the 3.1 release. We will re-introduce them in a future release.

* Polishing of nathan's PR

* New pdf

Co-authored-by: nathanfishcircleci <67339148+nathanfishcircleci@users.noreply.github.com>