SERVER-1081 | Edit object storage documentation for server 3 (#5399)
* Edit object storage documentation for server 3

S3 object storage is now slighly more general as we support S3
compatible storage like Minio. Also edited the GCS object storage docs
to make them similar to the S3 docs

* Server 3.0.2. changelog (#5398)

* 3.0.2. changelog

* Adding pdf

* periods. :/

* pdf

* Modifying blurb for 1074.

* New pdf

* Fixing changelog

* Server 3 changelog updates

* [MAC-564] Xcode 12.5 Release

* Remove 12.1.0 (12.1.1 is replacement)

* Maximum allowed contexts (#5402)

Introduced a threshold to the number of contexts allowed within the configuration

* Update caching.md (#5400)

example yaml was broken (wrong indent), so fix it.

* Changes to Server 3 Ops/Metrics Guide (#5403)

* Update server-3-install.adoc

Moved nomad mtls instructions down in the documentation as it was causing people to install nomad.

* Update server-3-operator-backup-and-restore.adoc

Add note that this isn't available yet.

* Update server-3-install.adoc

This section is really not ideal. I'd like to make the mTLS default to disabled so we can remove this entire section and tell the user to ignore till step 4. For now this is a slight improvement to provide more context.

* Adding links

* Update server-3-operator-metrics-and-monitoring.adoc

We removed LOKI and Grafana from the 3.1 release. We will re-introduce them in a future release.

* Polishing of nathan's PR

* New pdf

Co-authored-by: nathanfishcircleci <67339148+nathanfishcircleci@users.noreply.github.com>

* Update server-3-overview.adoc (#5386)

removed the loki and grafana related services

Co-authored-by: annabaker <anna.baker.313@gmail.com>

* Add remote Docker v20.10.6 to list (#5407)

* Add: table striped class to workflows document (#5408)

* polishing

Co-authored-by: Anna Baker <anna.baker.313@gmail.com>
Co-authored-by: Alexa <82475644+alexazl@users.noreply.github.com>
Co-authored-by: JadhavPoonam <poonam@circleci.com>
Co-authored-by: Yuki Inoue <inoueyuworks@gmail.com>
Co-authored-by: nathanfishcircleci <67339148+nathanfishcircleci@users.noreply.github.com>
Co-authored-by: Ricardo N Feliciano <FelicianoTech@gmail.com>
Co-authored-by: tees <12987958+teesloane@users.noreply.github.com>