Clarify local validation of files with private orb
I had to dig around for a definition of org-slug and it wasn't intuitive to me that the path was just the local config file (because it never needed to be specified before). 
It would also be wonderful if you could add what version of the orb the validation uses--the latest semver, maybe? I haven't been able to find that documented.