[WEB-610] Update long cookie expirations (#5991)
The goal of this ticket is to reduce our long (10-year) expirations on
various cookies to 2 years. This commit changes the expiry that we
write when creating cookies.

In addition, since cookie setting is conditional, we can't count on
those being re-created with the current expiry date on each visit. So
we add and call a function to update the expiration date on an existing
cookie (but won't set the cookie if it doesn't already exist).

See also: [WEB-611]

Co-authored-by: Romain Sylvian <romain@circleci.com>