[WEB-696] Split GTM noscript, move include to head
Per Google, they want the GTM tag as high in the head as possible, and
the noscript version in the body (also as high as possible).

I selected this position in the head somewhat arbitrarily, but it's
below the items that are actually critical to be at the beginning (meta
charset, redirect).