Fix ubuntu-2004:2022.04.1 image incorrectly states Ubuntu 22.04
Problem:
The Ubuntu Machine Executor image ubuntu-2004:2022.04.1 states that it is Ubuntu 22.04, but it is 20.04.
Confirmed via this job: https://app.circleci.com/pipelines/github/Chekote/BehatRetryExtension/18/workflows/7936a3df-4143-4b38-bfc6-eaba52ba8bc3/jobs/62

Cause:
Pull Request #6825 added additional images to the docs, and it seems that the lines were copy-and-pasted, but only the tag was updated on the newly pasted line. Unfortunately the Ubuntu version seems to have been overlooked.

Fix:
Correct the error so that the correct Ubuntu version is stated.