Upgrade minimum helm version to current latest
Earlier versions of helm (like 3.6) had `helm registry login`
as a beta feature. Upgrading to latest removes the need for
the `HELM_EXPERIMENTAL_OCI=1` variable. This is the error
you would receive on 3.6:

Error: this feature has been marked as experimental and is not enabled by default. Please set HELM_EXPERIMENTAL_OCI=1 in your environment to use this feature