Document collecting tests results for CMake / CTest projects
CMake is one of the most popular build systems for C and C++ projects.
Recently, I learnt that CTest can output JUnit test results that
CircleCI can ingest and we are using it for one of our internal projects
at Postman.

I wasn't able to run the docs locally, but I think these changes are a
good start for documenting how it works.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>