chore: Remove DNS suggestions
SERVER-2215

The DNS documentation can apparently be confusing for folks
that are not using GCP or AWS as their DNS provider. Instead
of trying to enumerate all the possible methods and providers
of managing DNS, we are going to remove it and assume a given
organization has a process to follow.