API-529: Public documentation for additional rate limit headers
With the move to the API Gateway, there are new headers that express the
functioning of the system. These `X-RateLimit-*` headers have been
added, as well as some clarifications around the other headers.

Note that the API Gateway adds both style headers, but upstream services
may still implement their own rate limiting and override the original
headers. A note has been added for this purpose.