Temporarily disable most types of files in Crowdin config
In an effort to get our Crowdin workflow flowing again, we decided to
start over from a minimal config and slowly build it back up.

This change disables all but a few simple files in the config. This
should let us get the workflow functioning again with a small subset of
the backlog of translation changes, and we can slowly and selectively
re-enable files/groups-of-files from here.