Add `update_as_unapproved` option to Crowdin config
This configuration option should allow stale translations to persist
until they're updated, rather than reverting to the fallback language.
This is the config we use on the circleci.com project.

One of the issues we're seeing is that of good Japanese translations
being reverted to English in the Crowdin PR, as Crowdin invalidates
those strings even for a trivial (i.e. typo fix) change.

Crowdin Docs:
https://developer.crowdin.com/configuration-file/#changed-strings-update