Define correct attribute mapping for OIDC GCP
The current docs are pretty vague about which attribute mappings are needed for the OIDC provider in GCP. The ones provided as example don't work (they can't even be saved because one of them seems to be invalid). This makes this already complex setup process even harder and it took me quite some googling and trial and error to figure out a mapping that actually works with the provided example config-file.
I hope this is helpful to others.