Add `update_as_unapproved` option to all items in Crowdin config
This follows up on a change first made in [1], which was inadvertently
not included in future updates to this config file. This commit adds
that option back to all items in the config, as originally intended.

The message on the previous commit describes the problem and intended
solution in more detail:

    This configuration option should allow stale translations to
    persist until they're updated, rather than reverting to the
    fallback language. This is the config we use on the circleci.com
    project.

    One of the issues we're seeing is that of good Japanese translations
    being reverted to English in the Crowdin PR, as Crowdin invalidates
    those strings even for a trivial (i.e. typo fix) change.

    Crowdin Docs:
        https://developer.crowdin.com/configuration-file/#changed-strings-update

[1] dc7deb7fc315148a4a74472a60af94fe35d64fe3