VSCode Extension - Clarify what qualifies a project to be selectable in setup
People who are less familiar with CircleCI and just want to see build statuses in their IDE can be confused as to why none of their projects show in the settings page, especially if none are autodetected. Reiterating the prerequsite of "A project you follow that is building on CircleCI" within the "Set up your project" section can drastically help remove cognitive overhead in trying to figure out what qualifies a project to show in this Project selection list. The point is very easy to miss otherwise.