Update config-intro.adoc
It seems like you had an earlier version of this tutorial that included lines for supplying Docker registry credentials and although these lines were taken out, the line numbers in the explanations that followed were not correct. 

I have attempted to correct them, and added in a small statement that notes the Docker credentials are omitted for your examples.