Fix Playwright Sharding command
The `npx playwright test -- --shard=${shard}` command has an extra `--` in it. When passing arguments to the command through `npm run test`, any arguments passed before the `--` will be interpreted by `npm`. But with `npx`, arguments are passed straight to the commande being run - in this instance, `playwright`, so the `--` is both unnecesary and causes the sharding to not work.