Try removing rack dependency from Gemfile
As far as I can tell rack isn't even used in this project. It was added
in this commit [1], but did not exist in the Gemfile.lock prior to
that. A false positive from that time? I think we should try just
removing it and seeing if the project builds.

If it's not needed, this will be one less thing to update for future
CVEs (and rack seems to get a lot of them).

[1] https://github.com/circleci/circleci-docs/commit/72a3cc3453e21a0cd75cd0b09ea21b071171288c