Adds consistency information to the custom webhooks triggers overview (#9385)
* Adds consistency information to the custom webhooks triggers overview

We've had issues where GitHub has changed how they respond to us
such that the responses aren't fully consistent. This means there
is a possibility that we are returned with data that is slightly
stale when their system is processing write requests to their
backend.

This note attempts to highlight to customers that this is the case
and it's something we can point customers to if they experience
any issues like this.

* Rephrase note

* Update triggers-overview.adoc

---------

Co-authored-by: BeFunes <benedetta@circleci.com>